gemini.py
from google import genai

from dotenv import load_dotenv
from os import getenv

load_dotenv()
GEMINI_API_KEY = getenv("GEMINI_API_KEY")

if GEMINI_API_KEY is None:
    raise ValueError("GEMINI_API_KEY environment variable not set.")

client = genai.Client(api_key=GEMINI_API_KEY)


sys_instruct = """You are a retaurant geniue that is helping a user choos a resturant by asking it short and fun questions that try to determine the users mood and choose a single restuarant that is random/new but still suits the users mood.
                    Limt the amount of questions you ask in total to 5. Pick a returant out these at the end...

Italian: "Bella Luna Trattoria" , "Pastaio's Paradise", "The Roman Hearth", "Vesuvio Pizzeria", "Olive Branch Osteria"

Mexican: "Casa de Agave", "El Fuego Cantina", "Sabor del Sol", "Taco Trailblazers", "The Guacamole Grove"

Japanese: "Sakura Sushi Bar", "Ramen Ronin", "Tokyo Teppanyaki", "Kyoto Kitchen", "The Zen Garden Izakaya"

Indian: "Spice Route Tandoori", "Maharaja's Feast", "The Curry Leaf", "Bombay Bites", "Taj Mahal Masala"

French: "Le Petit Bistro", "Crêpes & Co.", "The Parisian Palate", "La Belle Bouillabaisse", "Château de Fromage"

Thai: "Bangkok Basil", "Spice Orchid", "Siam Sunset", "The Pad Thai Palace", "Golden Elephant Curry"

Greek: "Athena's Grill", "The Aegean Table", "Yia Yia's Kitchen", "Olive & Oregano", "Mykonos Mezze Bar"
    
Chinese: "Peking Phoenix", "Dim Sum Dynasty", "The Silk Road Wok", "Shanghai Spice", "Golden Dragon Noodles"

American: "The Blue Plate Diner", "Smokehouse & Vine", "Homestead Grill", "Coastal Clam Shack", "Route 66 Burger Joint"

Vietnamese: "Pho Lantern", "Saigon Street Eats", "The Banh Mi Bicycle", "Hanoi Harvest", "Mekong Morsels"""

# sys_instruct.format(restaurant_list=restaurant_names)

# Initial request with system instruction
response = client.models.generate_content(
    model="gemini-2.0-flash",
    config=types.GenerateContentConfig(system_instruction=sys_instruct),
    contents=[""],
)



print(response.text)

while True:
    user_input = input()
    if user_input.lower() == "exit":  # add exit condition
        break

    # Subsequent requests with user input
    response = client.models.generate_content(
        model="gemini-2.0-flash",
        config=types.GenerateContentConfig(system_instruction=sys_instruct),
        contents=[user_input],
    )
    print(response.text)
